fixed wordlist size assert vs 64bit machine (hello glook)
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@1328 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -13508,7 +13508,7 @@ bool CSphIndex_VLN::SetupQueryWord ( CSphQueryWord & tWord, const CSphTermSetup 
 
 	if ( m_bPreloadWordlist )
 	{
-		assert ( uWordlistOffset>0 && uWordlistOffset<=(SphOffset_t)ULONG_MAX );
+		assert ( uWordlistOffset>0 && uWordlistOffset<=(SphOffset_t)UINT_MAX );
 		pBuf = &m_pWordlist[(DWORD)uWordlistOffset];
 	}
 	else
